+++ lib/subspace/cli.rb
@@ -33,8 +33,11 @@ class Subspace::Cli
       c.syntax = 'subspace init [vars]'
       c.summary = 'Run without options to initialize subspace.'
       c.description = 'Some initialization routines can be run indiviaully, useful for upgrading'
-      c.example 'init a new project', 'subspace init'
-      c.example 'create the new style application.yml vars template', 'subspace init vars'
+      c.example 'init a new project with one default environment (default dev)', 'subspace init'
+      c.example 'create a new fully automated production environment configuration', 'subspace init --terraform --env production'
+      c.option '--ansible', 'initialize ansible for managing individual servers'
+      c.option '--terraform', 'Initialize terraform for managing infrastructure'
+      c.option '--env STRING', 'Initialize configuration for a new environment'
       c.when_called Subspace::Commands::Init
     end