+++ lib/subspace/commands/init.rb
@@ -3,10 +3,20 @@ require 'erb'
 require 'securerandom'
 class Subspace::Commands::Init < Subspace::Commands::Base
   def initialize(args, options)
-    if args.first == "vars"
-      init_vars
-    else
-      run
+    options.default env: "dev"
+    @env = options[:env]
+
+    if !options[:ansible] && !options[:terraform]
+      # They didn't pass in any options (subspace init) so just do both
+      options[:ansibe] = options[:terraform] = true
+    end
+
+    if options[:ansible]
+      init_ansible
+    end
+
+    if options[:terraform]
+      init_terraform
     end
   end
 
@@ -16,36 +26,14 @@ class Subspace::Commands::Init < Subspace::Commands::Base
       abort unless answer.chomp == "yes"
     end
 
-    FileUtils.mkdir_p File.join dest_dir, "group_vars"
-    FileUtils.mkdir_p File.join dest_dir, "host_vars"
-    FileUtils.mkdir_p File.join dest_dir, "vars"
-    FileUtils.mkdir_p File.join dest_dir, "roles"
-
-    copy ".gitignore"
-    #template "../provision.rb"
-    template "ansible.cfg"
-    template "hosts"
-    template "group_vars/all"
-
-    create_vault_pass
-    environments.each do |env|
-      @env = env
-      @hostname = hostname(env)
-      template "group_vars/template", "group_vars/#{env}"
-      template "host_vars/template", "host_vars/#{env}"
-      create_vars_file_for_env env
-      template "playbook.yml", "#{env}.yml"
-    end
-    create_vars_file_for_env "development"
-    init_vars
+    init_ansible
 
     puts """
     1. Create a server.
 
     2. Set your server's location:
 
-      vim config/provision/host_vars/production
-      vim config/provision/host_vars/dev
+      vim config/subspace/inventory.yml
 
     3. Set up your authorized_keys:
 
@@ -54,25 +42,39 @@ class Subspace::Commands::Init < Subspace::Commands::Base
     4. Then provision your server:
 
       subspace provision dev
-      subspace provision production
 
   """
 
   end
 
-  def init_vars
-    FileUtils.mkdir_p File.join dest_dir, "templates"
-    copy "templates/application.yml.template"
+  private
+
+  def init_ansible
+    FileUtils.mkdir_p File.join dest_dir, "group_vars"
+    FileUtils.mkdir_p File.join dest_dir, "vars"
+    FileUtils.mkdir_p File.join dest_dir, "roles"
+
+    copy ".gitignore"
+    template "ansible.cfg"
+    template "inventory.yml", "inventory.#{@env}.yml"
+    template "group_vars/all"
+
+    create_vault_pass
+    @hostname = hostname(@env)
+    template "group_vars/template", "group_vars/#{@env}"
+    create_vars_file_for_env @env
+    template "playbook.yml", "#{@env}.yml"
+
+    create_vars_file_for_env "development" #TODO rename to local
+    init_appyml
   end
 
-  private
+  def init_terraform
+
+  end
 
   def project_name
     File.basename(Dir.pwd)
-  end
-
-  def environments
-    %w(production dev)
   end
 
   def hostname(env)
@@ -92,5 +94,10 @@ class Subspace::Commands::Init < Subspace::Commands::Base
     Dir.chdir dest_dir do
       `ansible-vault encrypt vars/#{env}.yml`
     end
+  end
+
+  def init_appyml
+    FileUtils.mkdir_p File.join dest_dir, "templates"
+    copy "templates/application.yml.template"
   end
 end