+++ lib/subspace/commands/provision.rb
@@ -9,7 +9,7 @@ class Subspace::Commands::Provision < Subspace::Commands::Base
   end
 
   def run
-    ansible_options = ["--diff"]
+    ansible_options = ["--diff", "-i", "inventory.#{@environment}.yml"]
     ansible_options = ansible_options | pass_through_params
     ansible_command "ansible-playbook", "#{@environment}.yml", *ansible_options
   end